move exitCode to sdk/common pkg
