+++ plugins/auth0/pkg/app/fetch.go
@@ -6,6 +6,7 @@ import (
 	"strings"
 
 	"github.com/aserto-dev/ds-load/plugins/auth0/pkg/httpclient"
+	"github.com/aserto-dev/ds-load/sdk/common"
 	"github.com/aserto-dev/ds-load/sdk/plugin"
 	"github.com/pkg/errors"
 
@@ -75,7 +76,7 @@ func (fetcher *FetchCmd) Fetch(results chan map[string]interface{}, errCh chan e
 		user, err := fetcher.readByPID()
 		if err != nil {
 			errCh <- err
-			SetExitCode(1)
+			common.SetExitCode(1)
 			return
 		}
 		results <- user
@@ -86,7 +87,7 @@ func (fetcher *FetchCmd) Fetch(results chan map[string]interface{}, errCh chan e
 		users, err := fetcher.readByEmail()
 		if err != nil {
 			errCh <- err
-			SetExitCode(1)
+			common.SetExitCode(1)
 			return
 		}
 		for _, user := range users {
@@ -107,7 +108,7 @@ func (fetcher *FetchCmd) Fetch(results chan map[string]interface{}, errCh chan e
 		ul, err := fetcher.mgmt.User.List(opts...)
 		if err != nil {
 			errCh <- err
-			SetExitCode(1)
+			common.SetExitCode(1)
 			return
 		}
 
@@ -115,21 +116,21 @@ func (fetcher *FetchCmd) Fetch(results chan map[string]interface{}, errCh chan e
 			res, err := u.MarshalJSON()
 			if err != nil {
 				errCh <- err
-				SetExitCode(1)
+				common.SetExitCode(1)
 				continue
 			}
 			var obj map[string]interface{}
 			err = json.Unmarshal(res, &obj)
 			if err != nil {
 				errCh <- err
-				SetExitCode(1)
+				common.SetExitCode(1)
 				continue
 			}
 			if fetcher.Roles {
 				roles, err := fetcher.getRoles(*u.ID)
 				if err != nil {
 					errCh <- err
-					SetExitCode(1)
+					common.SetExitCode(1)
 				} else {
 					obj["roles"] = roles
 				}
