+++ plugins/azuread/pkg/app/fetch.go
@@ -6,6 +6,7 @@ import (
 
 	"github.com/alecthomas/kong"
 	"github.com/aserto-dev/ds-load/plugins/azuread/pkg/azureclient"
+	"github.com/aserto-dev/ds-load/sdk/common"
 	"github.com/aserto-dev/ds-load/sdk/plugin"
 	kiota "github.com/microsoft/kiota-serialization-json-go"
 )
@@ -37,7 +38,7 @@ func Fetch(azureClient *azureclient.AzureADClient, results chan map[string]inter
 	aadUsers, err := azureClient.ListUsers()
 	if err != nil {
 		errors <- err
-		SetExitCode(1)
+		common.SetExitCode(1)
 	}
 
 	for _, user := range aadUsers.GetValue() {
@@ -45,13 +46,13 @@ func Fetch(azureClient *azureclient.AzureADClient, results chan map[string]inter
 		err := user.Serialize(writer)
 		if err != nil {
 			errors <- err
-			SetExitCode(1)
+			common.SetExitCode(1)
 			return
 		}
 		userBytes, err := writer.GetSerializedContent()
 		if err != nil {
 			errors <- err
-			SetExitCode(1)
+			common.SetExitCode(1)
 			return
 		}
 
@@ -60,7 +61,7 @@ func Fetch(azureClient *azureclient.AzureADClient, results chan map[string]inter
 		err = json.Unmarshal([]byte(userString), &obj)
 		if err != nil {
 			errors <- err
-			SetExitCode(1)
+			common.SetExitCode(1)
 			return
 		}
 		results <- obj
