+++ plugins/okta/cmd/ds-load-okta/main.go
@@ -5,6 +5,7 @@ import (
 
 	"github.com/alecthomas/kong"
 	"github.com/aserto-dev/ds-load/plugins/okta/pkg/app"
+	"github.com/aserto-dev/ds-load/sdk/common"
 	"github.com/aserto-dev/ds-load/sdk/common/kongyaml"
 )
 
@@ -34,5 +35,5 @@ func main() {
 	if err := ctx.Run(); err != nil {
 		ctx.FatalIfErrorf(err)
 	}
-	os.Exit(app.GetExitCode())
+	os.Exit(common.GetExitCode())
 }
