+++ plugins/okta/pkg/app/cli.go
@@ -2,23 +2,10 @@ package app
 
 import (
 	"fmt"
-	"sync/atomic"
 
 	"github.com/alecthomas/kong"
 	"github.com/aserto-dev/ds-load/sdk/common/version"
 )
-
-var (
-	exitCode int32
-)
-
-func GetExitCode() int {
-	return int(atomic.LoadInt32(&exitCode))
-}
-
-func SetExitCode(code int) {
-	atomic.StoreInt32(&exitCode, int32(code))
-}
 
 type CLI struct {
 	Config          kong.ConfigFlag    `help:"Configuration file path" short:"c"`
